feat(sdk): getStreamMetadataUrl
