+++ packages/sdk/src/riverConfig.ts
@@ -128,3 +128,18 @@ export function makeRiverConfig(inEnvironmentId?: string) {
     }
     return config
 }
+
+export const getStreamMetadataUrl = (environmentId: string) => {
+    switch (environmentId) {
+        case 'alpha':
+            return 'https://alpha.river.delivery'
+        case 'gamma':
+            return 'https://gamma.river.delivery'
+        case 'omega':
+            return 'https://river.delivery'
+        case 'delta':
+            return 'https://delta.river.delivery'
+        default:
+            return 'http://localhost:3002'
+    }
+}
